Still not optimal after 2 rewrites.

+++ SessyController/Configurations/SessyBatteryEndpoint.cs
@@ -9,8 +9,8 @@
         public string? UserId { get; set; }
         public string? Password { get; set; }
         public string? BaseUrl { get; set; }
-        public int MaxCharge { get; set; }
-        public int MaxDischarge { get; set; }
+        public double MaxCharge { get; set; }
+        public double MaxDischarge { get; set; }
         public double Capacity { get; set; }
     }
 }
